+++ gradle-plugin/kubernetes/src/test/java/org/eclipse/jkube/gradle/plugin/task/KubernetesBuildTaskTest.java
@@ -38,6 +38,7 @@ import org.mockito.MockedConstruction;
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 import static org.junit.Assert.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.mockConstruction;
@@ -64,7 +65,7 @@ public class KubernetesBuildTaskTest {
     dockerBuildServiceMockedConstruction = mockConstruction(DockerBuildService.class, (mock, ctx) -> {
       when(mock.isApplicable()).thenReturn(isBuildServiceApplicable);
       if (isBuildError) {
-        doThrow(new JKubeServiceException("Exception during Build")).when(mock).build(any());
+        doThrow(new JKubeServiceException("Exception during Build")).when(mock).build(anyList());
       }
     });
     isBuildServiceApplicable = true;
@@ -94,18 +95,7 @@ public class KubernetesBuildTaskTest {
     // Then
     assertThat(buildTask.jKubeServiceHub.getBuildService()).isNotNull()
         .isInstanceOf(DockerBuildService.class);
-    verify(buildTask.jKubeServiceHub.getBuildService(), times(1)).build(any());
-  }
-
-  @Test
-  public void runTask_withEmptyImageConfigurations_shouldNotRunBuild() throws JKubeServiceException {
-    // Given
-    extension.images = Collections.emptyList();
-    final KubernetesBuildTask buildTask = new KubernetesBuildTask(KubernetesExtension.class);
-    // When
-    buildTask.runTask();
-    // Then
-    verify(buildTask.jKubeServiceHub.getBuildService(), times(0)).build(any());
+    verify(buildTask.jKubeServiceHub.getBuildService(), times(1)).build(anyList());
   }
 
   @Test
@@ -149,6 +139,6 @@ public class KubernetesBuildTaskTest {
 
     // Then
     Assertions.assertThat(dockerBuildServiceMockedConstruction.constructed()).isEmpty();
-    verify(buildTask.jKubeServiceHub.getBuildService(), times(0)).build(any());
+    verify(buildTask.jKubeServiceHub.getBuildService(), times(0)).build(anyList());
   }
 }