+++ jkube-kit/config/service/src/test/java/org/eclipse/jkube/kit/config/service/openshift/OpenshiftBuildServiceIntegrationTest.java
@@ -15,6 +15,7 @@ package org.eclipse.jkube.kit.config.service.openshift;
 
 import java.io.File;
 import java.nio.charset.StandardCharsets;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
@@ -192,7 +193,7 @@ public class OpenshiftBuildServiceIntegrationTest {
     final BuildServiceConfig config = withBuildServiceConfig(defaultConfig.build());
     final WebServerEventCollector collector = prepareMockServer(config, true, false, false);
 
-    new OpenshiftBuildService(jKubeServiceHub).build(image);
+    new OpenshiftBuildService(jKubeServiceHub).build(Collections.singletonList(image));
 
     // we should find a better way to assert that a certain call has been made
     assertTrue(mockServer.getOpenShiftMockServer().getRequestCount() > 8);
@@ -214,7 +215,7 @@ public class OpenshiftBuildServiceIntegrationTest {
             .build())
         .build());
     // When
-    new OpenshiftBuildService(jKubeServiceHub).build(image);
+    new OpenshiftBuildService(jKubeServiceHub).build(Collections.singletonList(image));
     // Then
     collector.assertEventsRecordedInOrder("build-config-check", "new-build-config", "pushed");
   }
@@ -232,7 +233,7 @@ public class OpenshiftBuildServiceIntegrationTest {
     final OpenshiftBuildService openshiftBuildService = new OpenshiftBuildService(jKubeServiceHub);
     // When
     final JKubeServiceException result = assertThrows(JKubeServiceException.class, () ->
-        openshiftBuildService.build(image));
+        openshiftBuildService.build(Collections.singletonList(image)));
     // Then
     assertThat(result)
         .getCause().hasMessage("This image has already been flattened, you can only flatten the image once");
@@ -254,7 +255,7 @@ public class OpenshiftBuildServiceIntegrationTest {
     final WebServerEventCollector collector = prepareMockServer(withBuildServiceConfig(defaultConfig.build()),
         true, false, false);
     // When
-    new OpenshiftBuildService(jKubeServiceHub).build(image);
+    new OpenshiftBuildService(jKubeServiceHub).build(Collections.singletonList(image));
     // Then
     collector.assertEventsRecordedInOrder("build-config-check", "new-build-config", "pushed");
   }
@@ -265,7 +266,7 @@ public class OpenshiftBuildServiceIntegrationTest {
     final WebServerEventCollector collector = prepareMockServer(
         config, true, false, false);
 
-    new OpenshiftBuildService(jKubeServiceHub).build(image);
+    new OpenshiftBuildService(jKubeServiceHub).build(Collections.singletonList(image));
 
     // we should find a better way to assert that a certain call has been made
     assertTrue(mockServer.getOpenShiftMockServer().getRequestCount() > 8);
@@ -284,7 +285,7 @@ public class OpenshiftBuildServiceIntegrationTest {
         .resourceConfig(mockedResourceConfig).build());
     final WebServerEventCollector collector = prepareMockServer(dockerConfig, true, false, false);
 
-    new OpenshiftBuildService(jKubeServiceHub).build(image);
+    new OpenshiftBuildService(jKubeServiceHub).build(Collections.singletonList(image));
 
     assertTrue(mockServer.getOpenShiftMockServer().getRequestCount() > 8);
     collector.assertEventsRecordedInOrder("build-config-check", "new-build-config", "pushed");
@@ -304,7 +305,7 @@ public class OpenshiftBuildServiceIntegrationTest {
     final WebServerEventCollector collector = prepareMockServer("component1-component2-name",
         dockerConfig, true, false, false);
 
-    new OpenshiftBuildService(jKubeServiceHub).build(image);
+    new OpenshiftBuildService(jKubeServiceHub).build(Collections.singletonList(image));
 
     assertTrue(mockServer.getOpenShiftMockServer().getRequestCount() > 8);
     collector.assertEventsRecordedInOrder("build-config-check", "new-build-config", "pushed");
@@ -322,7 +323,7 @@ public class OpenshiftBuildServiceIntegrationTest {
         .build());
     final WebServerEventCollector collector = prepareMockServer(dockerConfig, true, false, false);
 
-    new OpenshiftBuildService(jKubeServiceHub).build(image);
+    new OpenshiftBuildService(jKubeServiceHub).build(Collections.singletonList(image));
 
     assertTrue(mockServer.getOpenShiftMockServer().getRequestCount() > 8);
     collector.assertEventsRecordedInOrder("build-config-check", "new-build-config", "pushed");
@@ -353,7 +354,7 @@ public class OpenshiftBuildServiceIntegrationTest {
             .build()
         ).build();
 
-    service.build(fromExtImage);
+    service.build(Collections.singletonList(fromExtImage));
 
     assertTrue(mockServer.getOpenShiftMockServer().getRequestCount() > 8);
     collector.assertEventsRecordedInOrder("build-config-check", "new-build-config", "pushed");
@@ -366,7 +367,7 @@ public class OpenshiftBuildServiceIntegrationTest {
     final BuildServiceConfig config = withBuildServiceConfig(defaultConfigSecret.build());
     final WebServerEventCollector collector = prepareMockServer(config, true, false, false);
 
-    new OpenshiftBuildService(jKubeServiceHub).build(image);
+    new OpenshiftBuildService(jKubeServiceHub).build(Collections.singletonList(image));
 
     // we should find a better way to assert that a certain call has been made
     assertTrue(mockServer.getOpenShiftMockServer().getRequestCount() > 8);
@@ -380,7 +381,7 @@ public class OpenshiftBuildServiceIntegrationTest {
     final OpenshiftBuildService openshiftBuildService = new OpenshiftBuildService(jKubeServiceHub);
 
     final JKubeServiceException result = assertThrows(JKubeServiceException.class, () ->
-        openshiftBuildService.build(image));
+        openshiftBuildService.build(Collections.singletonList(image)));
 
     assertThat(result).hasMessageContaining("Unable to build the image using the OpenShift build service");
   }
@@ -390,7 +391,7 @@ public class OpenshiftBuildServiceIntegrationTest {
     final BuildServiceConfig config = withBuildServiceConfig(defaultConfig.build());
     final WebServerEventCollector collector = prepareMockServer(config, true, true, true);
 
-    new OpenshiftBuildService(jKubeServiceHub).build(image);
+    new OpenshiftBuildService(jKubeServiceHub).build(Collections.singletonList(image));
 
     collector.assertEventsRecordedInOrder("build-config-check", "patch-build-config", "pushed");
     collector.assertEventsNotRecorded("new-build-config");
@@ -406,7 +407,7 @@ public class OpenshiftBuildServiceIntegrationTest {
         .resourceConfig(mockedResourceConfig).build());
     final WebServerEventCollector collector = prepareMockServer(config, true, false, false);
 
-    new OpenshiftBuildService(jKubeServiceHub).build(image);
+    new OpenshiftBuildService(jKubeServiceHub).build(Collections.singletonList(image));
 
     collector.assertEventsRecordedInOrder("build-config-check", "new-build-config", "pushed");
     collector.assertEventsNotRecorded("patch-build-config");
@@ -430,7 +431,7 @@ public class OpenshiftBuildServiceIntegrationTest {
     final BuildServiceConfig config = withBuildServiceConfig(dockerImageConfig.build());
     final WebServerEventCollector collector = prepareMockServer(config, true, false, false);
 
-    new OpenshiftBuildService(jKubeServiceHub).build(image);
+    new OpenshiftBuildService(jKubeServiceHub).build(Collections.singletonList(image));
 
     // we should add a better way to assert that a certain call has been made
     assertTrue(mockServer.getOpenShiftMockServer().getRequestCount() > 7);
@@ -445,7 +446,7 @@ public class OpenshiftBuildServiceIntegrationTest {
     final BuildServiceConfig config = withBuildServiceConfig(dockerImageConfigSecret.build());
     final WebServerEventCollector collector = prepareMockServer(config, true, false, false);
 
-    new OpenshiftBuildService(jKubeServiceHub).build(image);
+    new OpenshiftBuildService(jKubeServiceHub).build(Collections.singletonList(image));
 
     // we should find a better way to assert that a certain call has been made
     assertTrue(mockServer.getOpenShiftMockServer().getRequestCount() > 7);